feat: merge the api of application/agent from extend to instance
