+++ lib/egg.js
@@ -3,9 +3,11 @@
 const path = require('path');
 const fs = require('fs');
 const EggCore = require('egg-core').EggCore;
+const cluster = require('cluster-client');
 const Messenger = require('./core/messenger');
 const createHttpClient = require('./core/httpclient');
 const createLoggers = require('./core/logger');
+const Singleton = require('./core/singleton');
 
 const HTTPCLIENT = Symbol('EggApplication#httpclient');
 const LOGGERS = Symbol('EggApplication#loggers');
@@ -274,6 +276,50 @@ class EggApplication extends EggCore {
     process.removeListener('unhandledRejection', this._unhandledRejectionHandler);
     return super.close();
   }
+
+  /**
+   * 创建一个单例并添加到 app/agent 上
+   * @param {String} name 单例的唯一名字
+   * @param {Object} create - 单例的创建方法
+   */
+  addSingleton(name, create) {
+    const options = {};
+    options.name = name;
+    options.create = create;
+    options.app = this;
+    const singleton = new Singleton(options);
+    singleton.init();
+  }
+
+
+  /**
+   * Wrap the Client with Leader/Follower Pattern
+   *
+   * @description almost the same as Agent.cluster API, the only different is that this method create Follower.
+   *
+   * @see https://github.com/node-modules/cluster-client
+   * @param {Function} clientClass - client class function
+   * @param {Object} [options]
+   *   - {Boolean} [autoGenerate] - whether generate delegate rule automatically, default is true
+   *   - {Function} [formatKey] - a method to tranform the subscription info into a string，default is JSON.stringify
+   *   - {Object} [transcode|JSON.stringify/parse]
+   *     - {Function} encode - custom serialize method
+   *     - {Function} decode - custom deserialize method
+   *   - {Boolean} [isBroadcast] - whether broadcast subscrption result to all followers or just one, default is true
+   *   - {Number} [responseTimeout] - response timeout, default is 3 seconds
+   *   - {Number} [maxWaitTime|30000] - leader startup max time, default is 30 seconds
+   * @return {ClientWrapper} wrapper
+   */
+  cluster(clientClass, options) {
+    options = options || {};
+    // master 启动的时候随机分配的一个端口，保证在一台机器上不冲突
+    options.port = this._options.clusterPort;
+    // agent worker is leader, app workers are follower
+    options.isLeader = this.type === 'agent';
+    options.logger = this.coreLogger;
+    return cluster(clientClass, options);
+  }
+
 }
 
 module.exports = EggApplication;
